+++ idp_solver/Idp_to_Z3.py
@@ -24,16 +24,73 @@ TODO: vocabulary
 """
 
 
-from z3 import Or, Not, And, ForAll, Exists, Z3Exception, Sum, If
+from fractions import Fraction
+from z3 import (Or, Not, And, ForAll, Exists, Z3Exception, Sum, If, FreshConst,
+                Q, DatatypeRef, Const, BoolSort, IntSort, RealSort, Function,
+                EnumSort, BoolVal)
 
-from textx import get_location
-
+from idp_solver.Parse import ConstructedTypeDeclaration, RangeDeclaration, SymbolDeclaration
 from idp_solver.Expression import (Constructor, Expression, IfExpr,
                                    AQuantification, BinaryOperator,
                                    ADisjunction, AConjunction, AComparison,
                                    AUnary, AAggregate, AppliedSymbol,
                                    UnappliedSymbol, Number, Brackets,
-                                   Variable, TRUE, DSLException)
+                                   Variable, TRUE)
+from idp_solver.utils import BOOL, INT, REAL
+
+
+# class ConstructedTypeDeclaration  ###########################################################
+
+def translate(self):
+    if self.translated is None:
+        if self.name == BOOL:
+            self.translated = BoolSort()
+            self.constructors[0].type = BOOL
+            self.constructors[1].type = BOOL
+            self.constructors[0].translated = BoolVal(True)
+            self.constructors[1].translated = BoolVal(False)
+            self.constructors[0].py_value = True
+            self.constructors[1].py_value = False
+        else:
+            self.translated, cstrs = EnumSort(self.name, [c.name for c in
+                                                        self.constructors])
+            self.check(len(self.constructors) == len(cstrs), "Internal error")
+            for c, c3 in zip(self.constructors, cstrs):
+                c.translated = c3
+                c.py_value = c3
+                self.map[str(c)] = c
+    return self.translated
+ConstructedTypeDeclaration.translate = translate
+
+
+# class RangeDeclaration  ###########################################################
+
+def translate(self):
+    if self.translated is None:
+        if self.type == INT:
+            self.translated = IntSort()
+        else:
+            self.translated = RealSort()
+    return self.translated
+RangeDeclaration.translate = translate
+
+
+# class SymbolDeclaration  ###########################################################
+
+def translate(self):
+    if self.translated is None:
+        if len(self.sorts) == 0:
+            self.translated = Const(self.name, self.out.translate())
+        else:
+
+            if self.out.name == BOOL:
+                types = [x.translate() for x in self.sorts]
+                self.translated = Function(self.name, types + [BoolSort()])
+            else:
+                types = [x.translate() for x in self.sorts] + [self.out.translate()]
+                self.translated = Function(self.name, types)
+    return self.translated
+SymbolDeclaration.translate = translate
 
 
 # class Expression  ###########################################################
@@ -44,16 +101,20 @@ def translate(self):
     if self.simpler is not None:
         return self.simpler.translate()
     return self.translate1()
-
 Expression.translate = translate
+
+def reified(self) -> DatatypeRef:
+    if self._reified is None:
+        self._reified = Const(b'*'+self.code.encode(), BoolSort())
+        Expression.COUNT += 1
+    return self._reified
+Expression.reified = reified
 
 
 # class Constructor  ##########################################################
 
 def translate(self):
     return self.translated
-
-
 Constructor.translate = translate
 
 
@@ -63,8 +124,6 @@ def translate1(self):
     return If(self.sub_exprs[IfExpr.IF].translate(),
               self.sub_exprs[IfExpr.THEN].translate(),
               self.sub_exprs[IfExpr.ELSE].translate())
-
-
 IfExpr.translate1 = translate1
 
 
@@ -75,7 +134,9 @@ def translate1(self):
     if not self.q_vars:
         return self.sub_exprs[0].translate()
     else:
-        finalvars = [v.translate() for v in self.q_vars.values()]
+        for v in self.q_vars.values():
+            v.translated = FreshConst(v.sort.decl.translate())
+        finalvars = [v.translated for v in self.q_vars.values()]
         forms = [f.translate() for f in self.sub_exprs]
 
         if self.q == '∀':
@@ -85,8 +146,6 @@ def translate1(self):
             forms = Or(forms) if 1 < len(forms) else forms[0]
             forms = Exists(finalvars, forms)
         return forms
-
-
 AQuantification.translate1 = translate1
 
 
@@ -122,8 +181,6 @@ def translate1(self):
         except Exception as e:
             raise e
     return out
-
-
 BinaryOperator.translate1 = translate1
 
 
@@ -135,8 +192,6 @@ def translate1(self):
     else:
         out = Or([e.translate() for e in self.sub_exprs])
     return out
-
-
 ADisjunction.translate1 = translate1
 
 
@@ -148,8 +203,6 @@ def translate1(self):
     else:
         out = And([e.translate() for e in self.sub_exprs])
     return out
-
-
 AConjunction.translate1 = translate1
 
 
@@ -168,14 +221,12 @@ def translate1(self):
         try:
             out = out + [function(x, y)]
         except Z3Exception:
-            raise DSLException("{}{}{}".format(str(x),
-                               self.operator[i - 1], str(y)))
+            self.check(False,
+                       "{}{}{}".format(str(x), self.operator[i - 1], str(y)))
     if 1 < len(out):
         return And(out)
     else:
         return out[0]
-
-
 AComparison.translate1 = translate1
 
 
@@ -191,9 +242,7 @@ def translate1(self):
         function = AUnary.MAP[self.operator]
         return function(out)
     except:
-        raise self.create_error(f"Incorrect syntax {self}")
-
-
+        self.check(False, f"Incorrect syntax {self}")
 AUnary.translate1 = translate1
 
 
@@ -202,8 +251,6 @@ AUnary.translate1 = translate1
 def translate1(self):
     assert self.quantifier_is_expanded, "Cannot expand {self.code}"
     return Sum([f.translate() for f in self.sub_exprs])
-
-
 AAggregate.translate1 = translate1
 
 
@@ -217,9 +264,8 @@ def translate1(self):
         return If(arg >= 0, arg, -arg)
     else:
         try:
-            if len(self.sub_exprs) != self.decl.arity:
-                raise self.create_error(f"Incorrect number of arguments for"
-                                        f" {self}")
+            self.check(len(self.sub_exprs) == self.decl.arity,
+                       f"Incorrect number of arguments for {self}")
             if len(self.sub_exprs) == 0:
                 return self.decl.translate()
             else:
@@ -229,13 +275,11 @@ def translate1(self):
         except AttributeError as e:
             # Using argument on symbol that has no arity.
             if str(e) == "'RangeDeclaration' object has no attribute 'arity'":
-                raise self.create_error(f"Symbol {self} does not accept an"
-                                        f" argument")
+                self.check(False,
+                           f"Symbol {self} does not accept an argument")
             # Unknown error.
             else:
                 raise AttributeError(e)
-
-
 AppliedSymbol.translate1 = translate1
 
 
@@ -243,8 +287,6 @@ AppliedSymbol.translate1 = translate1
 
 def translate1(self):
     assert False, "Internal error"
-
-
 UnappliedSymbol.translate1 = translate
 
 
@@ -252,17 +294,32 @@ UnappliedSymbol.translate1 = translate
 
 def translate(self):
     return self.translated
-
-
 Variable.translate = translate
 
 
 # Class Number  #######################################################
 
 def translate(self):
+    if self.translated is None:
+        ops = self.number.split("/")
+        if len(ops) == 2:  # possible with str_to_IDP on Z3 value
+            self.py_value = Fraction(self.number)
+            self.translated = Q(self.py_value.numerator, self.py_value.denominator)
+            self.type = REAL
+        elif '.' in self.number:
+            v = self.number if not self.number.endswith('?') else self.number[:-1]
+            if "e" in v:
+                self.py_value = float(eval(v))
+                self.translated = self.py_value
+            else:
+                self.py_value = Fraction(v)
+                self.translated = Q(self.py_value.numerator, self.py_value.denominator)
+            self.type = REAL
+        else:
+            self.py_value = int(self.number)
+            self.translated = self.py_value
+            self.type = INT
     return self.translated
-
-
 Number.translate = translate
 
 
@@ -270,8 +327,6 @@ Number.translate = translate
 
 def translate1(self):
     return self.sub_exprs[0].translate()
-
-
 Brackets.translate1 = translate1